Code refactor
Separated various parts of the code which was a single file into separate files in a meaningful manner. Handled errors properly by propagating errors up the call stack and handling them in the calling function. Added CSV parser and data validators as middleware to the upload route. Further split the middleware file based on individual functionalities. Implemented proper folder and file structure.
