+++ package.json
@@ -16,6 +16,7 @@
     "decimal.js": "^10.4.3",
     "dotenv": "^16.4.1",
     "express": "^4.18.2",
+    "express-validator": "^7.0.1",
     "moment": "^2.30.1",
     "mongodb": "^6.3.0",
     "mongoose": "^8.1.0",
