+++ src/middleware/useMulter.js
@@ -1,0 +1,6 @@
+const multer = require("multer");
+
+const storage = multer.memoryStorage();
+const upload = multer({ storage: storage });
+
+module.exports = { upload };
