+++ src/views/mongoSchema.js
@@ -1,0 +1,16 @@
+const mongoose = require("mongoose");
+
+const transactionSchema = new mongoose.Schema(
+  {
+    Date: Date,
+    Description: String,
+    Amount: mongoose.Types.Decimal128,
+    Currency: String,
+    amountInINR: mongoose.Types.Decimal128,
+  },
+  { versionKey: false }
+);
+
+const Transaction = mongoose.model("Transaction", transactionSchema);
+
+module.exports = Transaction;
